+++ apps/api_admin/resources/code.py
@@ -1,10 +1,10 @@
 from fastapi import APIRouter, BackgroundTasks
 
-from extensions import Route, error_schema, resp_success
+from extensions import Route, ErrorSchema
 from apps.api_admin.entities import CreateCodeSchema, CreateCodeParser
 from apps.api_admin.logics import create_sms_code, send_sms_task
 
-router = APIRouter(route_class=Route, responses=error_schema)
+router = APIRouter(route_class=Route, responses=ErrorSchema)
 
 
 @router.post('', response_model=CreateCodeSchema, status_code=201)
@@ -18,4 +18,4 @@ async def create_code(parser: CreateCodeParser, background_tasks: BackgroundTask
 
     # background_tasks.add_task(send_sms_task, params=params, code=code)
 
-    return resp_success(data={'success': True, 'code': code})
+    return CreateCodeSchema(data={'success': True, 'code': code})