+++ apps/api_admin/resources/user.py
@@ -1,13 +1,13 @@
 from fastapi import APIRouter, Depends, Path
 
-from extensions import Route, Pagination, error_schema, resp_success
+from extensions import Route, Pagination, ErrorSchema
 from apps.modules import current_admin_user
 from apps.api_admin.entities import (
     ReadUserSchema, ListUserSchema, UserSchema, PatchUserParser, FilterUserParser
 )
 from apps.api_admin.logics import UserResolver
 
-router = APIRouter(route_class=Route, responses=error_schema, dependencies=[current_admin_user])
+router = APIRouter(route_class=Route, responses=ErrorSchema, dependencies=[current_admin_user])
 
 
 @router.get('/{u_id}', response_model=ReadUserSchema, status_code=200)
@@ -17,7 +17,7 @@ async def read_user(
     """the api of read one user"""
 
     rt = await UserResolver.read_user(u_id)
-    return resp_success(data=rt)
+    return ReadUserSchema(data=rt)
 
 
 @router.patch('/{u_id}', response_model=UserSchema, status_code=201)
@@ -28,7 +28,7 @@ async def patch_user(
     """the api of update one user"""
 
     user = await UserResolver.patch_user(u_id, parser)
-    return resp_success(data=user)
+    return UserSchema(data=user)
 
 
 @router.get('', response_model=ListUserSchema, status_code=200)
@@ -42,4 +42,4 @@ async def list_users(
     query = Pagination(query, parser.page, parser.pagesize).items()
     result = await UserResolver.response_users(await query)
 
-    return resp_success(data={'total': total, 'users': result})
+    return ListUserSchema(data={'total': total, 'users': result})