+++ apps/api_test/resources.py
@@ -1,9 +1,9 @@
 from fastapi import APIRouter
 
-from extensions import Route, error_schema, resp_success
+from extensions import Route, ErrorSchema, resp_success
 from apps.api_test.logics import RedisTestResolver
 
-router = APIRouter(route_class=Route, responses=error_schema)
+router = APIRouter(route_class=Route, responses=ErrorSchema)
 
 
 @router.post('/redis')