+++ apps/libs/exception.py
@@ -11,12 +11,12 @@ from extensions.log import logger
 from extensions.exceptions import BaseHTTPException
 
 
-def log_message(request: Request, message: Any):
+def log_message(method: str, url: str, message: Any):
     """log message when catch exception"""
 
-    logger.error('start error'.center(60, '*'))
-    logger.error(f'{request.method} {request.url}')
-    logger.error(f'error is {message}')
+    logger.error('start error, this is'.center(60, '*'))
+    logger.error(f'{method} {url}')
+    logger.error(message)
     logger.error('end error'.center(60, '*'))
 
 
@@ -25,15 +25,15 @@ def register_exception(app: FastAPI):
     async def catch_c_http_exception(request: Request, exc: BaseHTTPException):
         """catch custom exception"""
 
-        log_message(request, exc.message)
-        content = {'code': exc.code, 'message': exc.message, 'data': None}
+        log_message(request.method, request.url, exc.message)
+        content = exc.response()
         return JSONResponse(content=content, status_code=exc.status_code, headers=exc.headers)
 
     @app.exception_handler(HTTPException)
     async def http_exception_handler(request: Request, exc: HTTPException):
         """catch FastAPI HTTPException"""
 
-        log_message(request, exc.detail)
+        log_message(request.method, request.url, exc.detail)
         content = {'code': StatusCode.bad_request, 'message': exc.detail, 'data': None}
         return JSONResponse(content=content, status_code=exc.status_code, headers=exc.headers)
 
@@ -41,8 +41,8 @@ def register_exception(app: FastAPI):
     async def assert_exception_handle(request: Request, exc: AssertionError):
         """catch Python AssertError"""
 
-        exc_str = ''.join(exc.args)
-        log_message(request, exc_str)
+        exc_str = ' '.join(exc.args)
+        log_message(request.method, request.url, exc_str)
         content = {'code': StatusCode.validator_error, 'message': exc_str, 'data': None}
         return JSONResponse(content=content, status_code=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
@@ -51,7 +51,7 @@ def register_exception(app: FastAPI):
         """catch tortoise-orm ValidatorError"""
 
         exc_str = '|'.join(exc.args)
-        log_message(request, exc_str)
+        log_message(request.method, request.url, exc.args)
         content = {'code': StatusCode.validator_error, 'message': exc_str, 'data': None}
         return JSONResponse(content=content, status_code=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
@@ -60,7 +60,7 @@ def register_exception(app: FastAPI):
         """catch FastAPI RequestValidationError"""
 
         exc_str = f'{exc}'.replace('\n', ' ').replace('   ', ' ')
-        log_message(request, exc_str)
+        log_message(request.method, request.url, exc)
         # content = exc.errors()
         content = {'code': StatusCode.validator_error, 'message': exc_str, 'data': None}
         return JSONResponse(content=content, status_code=status.HTTP_422_UNPROCESSABLE_ENTITY)
@@ -69,6 +69,6 @@ def register_exception(app: FastAPI):
     async def exception_handle(request: Request, exc: Exception):
         """catch other exception"""
 
-        log_message(request, traceback.format_exc())
+        log_message(request.method, request.url, traceback.format_exc())
         content = {'code': StatusCode.server_error, 'message': str(exc), 'data': None}
         return JSONResponse(content=content, status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)